backend/requirements.txt
fastapi
uvicorn[standard]

SQLAlchemy
psycopg[binary]

alembic
pydantic-settings
python-dotenv

python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1

python-multipart
